requirements.txt
fastapi==0.115.12
fitz==0.0.1.dev2
langchain_core==0.3.63
langchain_ollama==0.3.3
markdown2==2.5.3
pandas==2.2.3
psycopg2==2.9.10
pydantic==2.11.5
pypandoc==1.15
python_docx==1.1.2
uvicorn==0.34.3
weasyprint==65.1
